move helper function to models.py
